LayoutTests/editing/mac/spelling/disable-automatic-spelling-correction-context-menu-item-expected.txt



Verifies that autocorrect='off' disables the 'Correct Spelling Automatically' context menu item on macOS. To manually test, check that the menu item is disabled when right clicking the top text field, but enabled when clicking the bottom text field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isSpellingCorrectionMenuItemEnabled(topInput) is false
PASS isSpellingCorrectionMenuItemEnabled(bottomInput) is true
PASS successfullyParsed is true

TEST COMPLETE
